refactor: move cli utils
Move `linkml.utils.cli_utils` to `linkml.cli.logging`.
